WIP: refactoring shoppinglist
- begining to re-implement functionality from shopping list display to
the new types defined in the store
